% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{package}
\name{consensusMIBC-package}
\alias{consensusMIBC}
\alias{consensusMIBC-package}
\title{consensusMIBC}
\description{
This package implements a nearest-centroid transcriptomic classifier, that assigns class 
labels according to the consensus molecular classification of 
Muscle-Invasive Bladder Cancer. For more information, please read the package vignette by
calling `vignette("consensusMIBC")`.
}
\note{
This is a contribution from the Tumor Identity Cards (CIT) program founded by the 'Ligue Nationale Contre le Cancer' (France): 
\url{http://cit.ligue-cancer.net}. For any question please contact \url{CITR@ligue-cancer.net}
}
\seealso{
\link[consensusMIBC]{getConsensusClass}, \link[consensusMIBC]{plotCorrelations}
}
\author{
Aurelie Kamoun
}
