% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sjo.dat}
\alias{sjo.dat}
\title{Gene expression data derived from the Sjodahl et. al. (2017) cohort}
\format{An object of class \code{matrix} with 22228 rows and 243 columns.}
\usage{
sjo.dat
}
\description{
A matrix of gene expression values for 22228 ENTREZ gene IDs and 243
samples derived from the Sjodahl 2017 cohort.
}
\references{
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/path.4886}
}
\keyword{datasets}
