% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{tcga.dat}
\alias{tcga.dat}
\title{Gene expression data derived from RNA-seq of the TCGA-BLCA cohort}
\format{An object of class \code{matrix} with 17987 rows and 406 columns.}
\usage{
tcga.dat
}
\description{
A matrix of gene expression values for 17987 ENTREZ gene IDs and 406
samples derived from the TCGA-BLCA cohort.
}
\references{
\url{https://portal.gdc.cancer.gov/projects/TCGA-BLCA}
}
\keyword{datasets}
